+++ src/it/scala-2.12/org/apache/james/gatling/imap/ImapAuthenticationIT.scala
@@ -1,4 +1,4 @@
-package org.apache.james.gatling
+package org.apache.james.gatling.imap
 
 import org.apache.james.gatling.imap.scenari.ImapAuthenticationScenario
