+++ src/it/scala-2.12/org/apache/james/gatling/imap/ImapFullIT.scala
@@ -1,6 +1,8 @@
-package org.apache.james.gatling
+package org.apache.james.gatling.imap
 
+import org.apache.james.gatling.Fixture
 import org.apache.james.gatling.imap.scenari.ImapFullScenario
+
 import scala.concurrent.duration._
 
 class ImapFullIT extends ImapIT {
