+++ src/it/scala-2.12/org/apache/james/gatling/jmap/JmapAuthenticationIT.scala
@@ -1,4 +1,4 @@
-package org.apache.james.gatling
+package org.apache.james.gatling.jmap
 
 import org.apache.james.gatling.jmap.scenari.JmapAuthenticationScenario
