+++ src/it/scala-2.12/org/apache/james/gatling/jmap/JmapBigSetIT.scala
@@ -1,5 +1,6 @@
-package org.apache.james.gatling
+package org.apache.james.gatling.jmap
 
+import org.apache.james.gatling.Fixture
 import org.apache.james.gatling.jmap.scenari.JmapBigSetScenario
 
 import scala.concurrent.duration._
