+++ src/it/scala-2.12/org/apache/james/gatling/jmap/JmapIT.scala
@@ -1,4 +1,4 @@
-package org.apache.james.gatling
+package org.apache.james.gatling.jmap
 
 import io.gatling.core.Predef._
 import io.gatling.core.funspec.GatlingFunSpec
@@ -6,6 +6,7 @@ import io.gatling.core.protocol.Protocol
 import io.gatling.core.structure.ScenarioBuilder
 import io.gatling.http.Predef._
 import org.apache.james.gatling.Fixture.{bart, simpsonDomain}
+import org.apache.james.gatling.JamesServer
 import org.apache.james.gatling.JamesServer.RunningServer
 import org.apache.james.gatling.control.RecipientFeeder.RecipientFeederBuilder
 import org.apache.james.gatling.control.UserFeeder.UserFeederBuilder
